requirements.txt
celery==5.2.7
Django==4.1.6
django_environ==0.9.0
django-filter==22.1
django-nested-admin==4.0.2
djangorestframework==3.14.0
factory-boy==3.2.1
gunicorn==20.1.0
Pillow==9.4.0
psycopg2-binary==2.9.5
redis==4.5.1
